refactor(authentication): 로그인 성공 후 멤버 전역 상태 생성
