+++ frontend/src/hooks/authentication/useLoginForm.ts
@@ -1,18 +1,11 @@
 import { FormEvent } from 'react';
-import { useNavigate } from 'react-router-dom';
 
-import { postLogin } from '@/api/authentication';
-import { ToastContext } from '@/context/ToastContext';
+import { useLoginMutation } from '@/queries/authentication/useLoginMutation';
 import { useInputWithValidate } from '../useInputWithValidate';
-import useCustomContext from '../utils/useCustomContext';
-import { useAuth } from './useAuth';
 import { validateEmail, validatePassword } from './validates';
 
 export const useLoginForm = () => {
-  const navigate = useNavigate();
-  const { failAlert, successAlert } = useCustomContext(ToastContext);
-
-  const { handleLoginState } = useAuth();
+  const { mutateAsync } = useLoginMutation();
 
   const {
     value: email,
@@ -32,18 +25,7 @@ export const useLoginForm = () => {
     e.preventDefault();
 
     if (isFormValid()) {
-      const response = await postLogin({ email, password });
-
-      if (!response.ok) {
-        console.error(response);
-        failAlert('로그인에 실패하였습니다.');
-
-        return;
-      }
-
-      handleLoginState(true);
-      navigate('/');
-      successAlert('로그인 성공!');
+      await mutateAsync({ email, password });
     }
   };
